+++ minecraft/src/net/minecraft/command/EntitySelectorOptions.java
@@ -95,7 +95,7 @@ public class EntitySelectorOptions {
 			putOption("distance", entitySelectorReader -> {
 				int i = entitySelectorReader.getReader().getCursor();
 				NumberRange.FloatRange floatRange = NumberRange.FloatRange.parse(entitySelectorReader.getReader());
-				if ((floatRange.getMin() == null || !((Float)floatRange.getMin() < 0.0F)) && (floatRange.getMax() == null || !((Float)floatRange.getMax() < 0.0F))) {
+				if ((floatRange.getMin() == null || !((Double)floatRange.getMin() < 0.0)) && (floatRange.getMax() == null || !((Double)floatRange.getMax() < 0.0))) {
 					entitySelectorReader.setDistance(floatRange);
 					entitySelectorReader.setLocalWorldOnly();
 				} else {
